requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
crypto==1.4.1
Django==5.0.3
idna==3.6
Naked==0.1.32
pillow==10.2.0
pypng==0.20220715.0
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
shellescape==3.8.1
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
